move to main and re-add dir check
